requirements.txt
tensorflow==2.15.0
numpy>=1.24.0
grpcio>=1.62.0
protobuf>=4.25.0
matplotlib>=3.8.0
flask>=3.0.0
requests>=2.31.0
google-cloud-storage>=2.14.0
